// src/modules/SOL/enums/bid-status.enum.ts
export enum BidStatusEnum {
    awaiting = 'awaiting',
    deserted = 'deserted',
    draft = 'draft',
    analysis = 'analysis',
    released = 'released',
    open = 'open',
    completed = 'completed',
    reopened = 'reopened',
    failed = 'failed',
    canceled = 'canceled',
    tiebreaker = 'tiebreaker'
}